+++ src/com/dakoda/alr/entity/player/Player.java
@@ -1,4 +1,5 @@
-package com.dakoda.alr.entity;
+package com.dakoda.alr.entity.player;
+import com.dakoda.alr.entity.Entity;
 
 public final class Player extends Entity {
